(PC-12942)[PRO]: rename custom hooks
