+++ pro/src/components/hooks/useOfferStockEditionURL.ts
