+++ pro/src/components/pages/Offers/Offer/Confirmation/Confirmation.jsx
@@ -7,12 +7,13 @@ import React, { useCallback } from 'react'
 import { Redirect } from 'react-router'
 import { Link } from 'react-router-dom'
 
-import useOfferEditionURL from 'components/hooks/useFeatureFlaggedOfferEditionURL'
 import { ReactComponent as PendingIcon } from 'components/pages/Offers/Offer/Confirmation/assets/pending.svg'
 import { ReactComponent as ValidateIcon } from 'components/pages/Offers/Offer/Confirmation/assets/validate.svg'
 import OfferPreviewLink from 'components/pages/Offers/Offer/OfferPreviewLink/OfferPreviewLink'
 import { OFFER_STATUS_PENDING } from 'components/pages/Offers/Offers/_constants'
 import { queryParamsFromOfferer } from 'components/pages/Offers/utils/queryParamsFromOfferer'
+
+import useOfferEditionURL from '../../../../hooks/useOfferEditionURL'
 
 const Confirmation = ({ isCreatingOffer, location, offer, setOffer }) => {
   const editionUrl = useOfferEditionURL(offer.isEducational, offer.id)
