+++ pro/src/components/pages/Offers/Offer/Stocks/ThingStocks.jsx
@@ -8,7 +8,6 @@ import React, { Fragment, useCallback, useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
 import { v4 as generateRandomUuid } from 'uuid'
 
-import useOfferEditionURL from 'components/hooks/useFeatureFlaggedOfferEditionURL'
 import PageTitle from 'components/layout/PageTitle/PageTitle'
 import { isOfferDisabled } from 'components/pages/Offers/domain/isOfferDisabled'
 import OfferStatusBanner from 'components/pages/Offers/Offer/OfferDetails/OfferStatusBanner/OfferStatusBanner'
@@ -27,6 +26,7 @@ import { ReactComponent as AddStockSvg } from 'icons/ico-plus.svg'
 import * as pcapi from 'repository/pcapi/pcapi'
 import { SubmitButton } from 'ui-kit'
 
+import useOfferEditionURL from '../../../../hooks/useOfferEditionURL'
 import { queryParamsFromOfferer } from '../../utils/queryParamsFromOfferer'
 
 import PriceErrorHTMLNotification from './PriceErrorHTMLNotification'
