+++ pro/src/new_components/OfferBreadcrumb/OfferBreadcrumb.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 
-import useOfferEditionURL from 'components/hooks/useFeatureFlaggedOfferEditionURL'
-import useOfferStockEditionURL from 'components/hooks/useFeatureFlaggedOfferStockEditionURL'
 import Breadcrumb, {
   BreadcrumbStyle,
 } from 'new_components/Breadcrumb/Breadcrumb'
+
+import useOfferEditionURL from '../../components/hooks/useOfferEditionURL'
+import useOfferStockEditionURL from '../../components/hooks/useOfferStockEditionURL'
 
 export enum OfferBreadcrumbStep {
   DETAILS = 'details',
